Rename BulkApiContext.ts to main.ts to be similar tpo other packages, add comments.
