+++ packages/bulkapi/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./database/index.js"
 export * from "./controllers/index.js"
-export * from "./BulkApiContext.js"
+export * from "./main.js"
