+++ packages/dbadmin/src/main.ts
@@ -1,10 +1,14 @@
-import { runWithTry } from "@lionweb/repository-common";
 import { Express } from "express"
 import pgPromise from "pg-promise"
 import pg from "pg-promise/typescript/pg-subset.js"
+import { runWithTry } from "@lionweb/repository-common";
 import { DBAdminApi, DBAdminApiImpl } from "./controllers/DBAdminApi.js"
 import { DBAdminApiWorker } from "./database/DBAdminApiWorker.js"
 
+/**
+ * Object containing 'global' contextual objects for this API.
+ * Avoids using glocal variables, as they easily get mixed up between the various API packages.
+ */
 export class DbAdminApiContext {
     dbConnection: pgPromise.IDatabase<object, pg.IClient>
     pgp: pgPromise.IMain<object, pg.IClient>
@@ -19,15 +23,20 @@ export class DbAdminApiContext {
     }
 }
 
+/**
+ * Register all api methods with the _app_
+ * @param app           The app to which the api is registered
+ * @param dbConnection  The database connection to be used by this API
+ * @param pgp           The pg-promise object to gain access to the pg helpers
+ */
 export function registerDBAdmin(
-    app: Express, 
-    dbConnection: pgPromise.IDatabase<object, pg.IClient>, 
-    pgp: pgPromise.IMain<object, pg.IClient>) 
-{
+    app: Express,
+    dbConnection: pgPromise.IDatabase<object, pg.IClient>,
+    pgp: pgPromise.IMain<object, pg.IClient>) {
     console.log("Registering DB Admin Module");
     // Create all objects 
     const dbAdminApiContext = new DbAdminApiContext(dbConnection, pgp)
-    
+
     // Add routes to app
     app.post("/init", runWithTry(dbAdminApiContext.dbAdminApi.init))
 }
