+++ packages/inspection/src/main.ts
@@ -1,12 +1,16 @@
-import { runWithTry, } from "@lionweb/repository-common";
 import { Express } from "express"
 import pgPromise from "pg-promise"
 import pg from "pg-promise/typescript/pg-subset.js"
 
+import { runWithTry, } from "@lionweb/repository-common";
 import { createInspectionApiWorker, InspectionApiWorker } from "./database/InspectionApiWorker.js"
 import { createInspectionApi, InspectionApi } from "./controllers/InspectionApi.js"
 import { InspectionQueries } from "./database/InspectionQueries.js";
 
+/**
+ * Object containing 'global' contextual objects for this API.
+ * Avoids using glocal variables, as they easily get mixed up between the various API packages.
+ */
 export class InspectionContext {
     dbConnection: pgPromise.IDatabase<object, pg.IClient>
     pgp: pgPromise.IMain<object, pg.IClient>
@@ -22,6 +26,13 @@ export class InspectionContext {
         this.inspectionApiWorker = createInspectionApiWorker(this)
     }
 }
+
+/**
+ * Register all api methods with the _app_
+ * @param app           The app to which the api is registered
+ * @param dbConnection  The database connection to be used by this API
+ * @param pgp           The pg-promise object to gain access to the pg helpers
+ */
 
 export function registerInspection(app: Express, dbConnection: pgPromise.IDatabase<object, pg.IClient>, pgp: pgPromise.IMain<object, pg.IClient>) {
     console.log("Registering Inspection Module");
